colocate user & messages redux tests
