+++ frontend/src/reducers/__tests__/user.test.js
@@ -1,24 +1,5 @@
-import userReducer from "../reducers/user";
-import {
-  setNickname,
-  SET_NICKNAME,
-  REQUEST_SET_NICKNAME,
-  setUserId,
-  SET_USERID
-} from "../actions/user";
-
-describe("user action creators", () => {
-  it("setNickname", () => {
-    const nickname = "Eric";
-    const action = setNickname(nickname);
-    expect(action).toEqual({ type: REQUEST_SET_NICKNAME, nickname });
-  });
-  it("setUserId", () => {
-    const userId = 123;
-    const action = setUserId(userId);
-    expect(action).toEqual({ type: SET_USERID, id: userId });
-  });
-});
+import userReducer from "../user";
+import { setNickname, SET_NICKNAME, setUserId } from "../../actions/user";
 
 describe("user reducer", () => {
   const initialState = userReducer(undefined, {});
